+++ entity/src/entities/video.rs
@@ -8,7 +8,7 @@ pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
     pub favorite_id: i32,
-    pub upper_id: i32,
+    pub upper_id: i64,
     pub upper_name: String,
     pub upper_face: String,
     pub name: String,
